# Improve performance and correctness of memory callbacks