## Changes

- Implement a specialized collection for `IMemoryCallback` used in `MemoryCallbackSystem` with copy-on-write behavior to safely handle adding/removing callbacks from within a callback.
- Add some basic tests for `MemoryCallbackSystem`

635fff6c5d574b161129d52a61ccead2e8897b0d addressed #1823, but the logic is still not quite correct: when a callback is removed, it will skip whatever callback is last in the list rather than the one that was removed. It also causes a significant performance hit on something that can run countless times per frame, due to the allocations and (un)subscribing `CollectionChanged` (which can be surprisingly expensive on multicast delegates).

This PR replaces the use of `ObservableCollection<T>` with a custom, pared down collection that only handles the needs of  `MemoryCallbackSystem`. It uses copy-on-write while the collection is being enumerated to allow modifying callbacks but keep the enumeration allocation free.

Some very rough performance numbers for Gargoyles (Genesis):
| Script                                                     | FPS before PR | FPS after PR |
|------------------------------------------------------------|---------------|--------------|
| None                                                       | 1000          | 1000         |
| Dummy script (only frameadvance loop)                      | 700           | 700          |
| Dummy script plus `event.on_bus_exec(function() end, 0x0)` | 215           | 500          |
| Gargoyles.lua (bundled)                                    | 115           | 175          |

## Issues/limitations

- Adding/removing callbacks from within a memory callback can be slightly more expensive
- Adding/removing a callback from within a callback will only come into effect on the next read/write/execute
In previous implementations this behavior has been inconsistent


[//]: # "Apart from the mandatory license signature, these tasks are optional, but doing them could save reviewers some time and get the PR merged sooner."
Check if completed:
- [x] I have run any relevant test suites
- [x] I, the committer, have read the [licensing terms for contributors](https://github.com/TASEmulators/BizHawk/blob/master/contributing.md#copyrights-and-licensing) (last updated 2022-07-15) and am compliant