@kalimag 2022-12-09
1. Any item is removed from `cbs`
2. The `CollectionChanged` handler does `e--`
3. The `while` loop will exit earlier
5. The last item in `cbsCopy` is not processed, regardless of which item was removed from `cbs`.