@ItEndsWithTens 2023-02-03
I don't have the skillset to speak to the correctness of this PR, but in terms of performance I can confirm it's a night and day difference!

I'm working on [an external tool for the original Silent Hill](https://github.com/ItEndsWithTens/SilentHillMapExaminer) to help with reverse engineering, namely picking apart the map format. To help me visualize things I'm drawing a debug overlay with either the GUI API or, when available (i.e. in Octoshock), with a memory callback that's called when the game swaps its double buffers, letting me draw directly into the PSX's VRAM. Drawing right on the framebuffer keeps everything rock solid synced with the game content, and is less jarring than the swimmy disconnect otherwise seen when waiting for a PostFrame ToolUpdate call and rendering on top of the tearing. Not strictly necessary, but nice to have.

Preparing for BizHawk 2.9 I'm seeing a substantial perf drop for the same code, from 60fps down to the mid 30s or so, between 2.8 and the latest HEAD (currently b1e8d93f5f4d26198ee7be2d9af7f718b332a22f); git bisect suggests it's the fault of the aforementioned 635fff6. Checking out this pull request instead I'm right back up to full speed without breaking a sweat.

What my anecdotal evidence is worth I can't say, but I thought I should share. If you'd rather I open a proper bug report for the performance drop, I can do that too, but since this PR seems to have addressed the issue it seemed redundant to bring it up elsewhere.